Update `maybe_clear_terminal` to remove scrollback
`IO.ANSI.clear` + `home` seem to not clear scrollback (atleast on macOS), i.e. it prints "one screen height" of blank space and resets the cursor to the first position. A little bit of digging through [ANSI escape codes](https://askubuntu.com/questions/25077/how-to-really-clear-the-terminal) led me to this solution. I'm not sure if this is the best approach/if there's a way to get the right sequence out from the `IO.ANSI` library.